+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]## [0.5.4.5b4]
+## [Unreleased]## [0.5.4.5b5]
+
+## [0.5.4.5b4] - 2022-10-17
 ### Added
 - OptimizationEngine.create_do_model() to instantiate as CPOptimizer model
 - OptimizationEngine CPOptimizer methods to create dvar as DataFrame columns